chamado/serializers.py
from dataclasses import field, fields
from chamado.models import Chamado

from rest_framework import serializers

class ChamadoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Chamado
        fields = '__all__'